remove RelationalResolver class; fix tests; move classes in packages
