--- forsuredbcompiler/src/main/java/com/fsryan/forsuredb/annotationprocessor/generator/code/resolver/RelationalResolverGenerator.java
@@ -1,19 +1,0 @@
-package com.fsryan.forsuredb.annotationprocessor.generator.code.resolver;
-
-import com.fsryan.forsuredb.annotationprocessor.TableContext;
-import com.fsryan.forsuredb.api.RelationalResolver;
-import com.fsryan.forsuredb.api.info.TableInfo;
-import com.squareup.javapoet.ClassName;
-
-import javax.annotation.processing.ProcessingEnvironment;
-
-/*package*/ class RelationalResolverGenerator extends ResolverGenerator {
-    protected RelationalResolverGenerator(ProcessingEnvironment processingEnv, TableInfo table, TableContext targetContext) {
-        super(processingEnv, table, targetContext);
-    }
-
-    @Override
-    protected ClassName extendsFromClassName() {
-        return ClassName.get(RelationalResolver.class);
-    }
-}
